Restore Py2 compatibility
