+++ bindings/python/unicorn/unicorn_py3/__init__.py
@@ -1,0 +1,1 @@
+from .unicorn import *
